requirements.txt
Flask==1.1.1
python-telegram-bot==12.2.0
requests-oauthlib==1.3.0
